Moved expectation value calculation into utils folder
