+++ lightworks/tomography/state_tomography.py
@@ -18,9 +18,9 @@ from typing import Callable
 import numpy as np
 
 from ..sdk.circuit import Circuit
-from ..sdk.state import State
 from .mappings import MEASUREMENT_MAPPING, PAULI_MAPPING
 from .utils import (
+    _calculate_expectation_value,
     _get_required_tomo_measurements,
     _get_tomo_measurements,
     state_fidelity,
@@ -160,9 +160,7 @@ class StateTomography:
             for c in _get_tomo_measurements(self.n_qubits)
         }
 
-        self._rho = StateTomography._calculate_density_matrix(
-            self.n_qubits, results_dict
-        )
+        self._rho = self._calculate_density_matrix(results_dict)
         return self.rho
 
     def fidelity(self, rho_exp: np.ndarray) -> float:
@@ -211,18 +209,12 @@ class StateTomography:
 
         return circuit
 
-    @staticmethod
-    def _calculate_density_matrix(
-        n_qubits: int, results: dict[str, dict]
-    ) -> np.ndarray:
+    def _calculate_density_matrix(self, results: dict[str, dict]) -> np.ndarray:
         """
         Calculates the density matrix using a provided dictionary of results
         data.
 
         Args:
-
-            n_qubits (int) : The number of dual-rail encoded qubits to calculate
-                the density matrix for.
 
             results (Result | dict) : Result containing measured output states
                 and counts. The keys of this dictionary should be the
@@ -237,12 +229,10 @@ class StateTomography:
 
         """
         # Process results to find density matrix
-        rho = np.zeros((2**n_qubits, 2**n_qubits), dtype=complex)
+        rho = np.zeros((2**self.n_qubits, 2**self.n_qubits), dtype=complex)
         for measurement, result in results.items():
-            expectation = StateTomography._calculate_expectation_value(
-                measurement, result
-            )
-            expectation /= 2**n_qubits
+            expectation = _calculate_expectation_value(measurement, result)
+            expectation /= 2**self.n_qubits
             # Calculate tensor product of the operators used
             ops = measurement.split(",")
             mat = PAULI_MAPPING[ops[0]]
@@ -251,44 +241,3 @@ class StateTomography:
             # Updated density matrix
             rho += expectation * mat
         return rho
-
-    @staticmethod
-    def _calculate_expectation_value(
-        measurement: str, results: dict[str, int]
-    ) -> float:
-        """
-        Calculates the expectation value for a given measurement and set of
-        results.
-
-        Args:
-
-            measurement (str) : The measurement operator used for the
-                computation.
-
-            results (dict) : A dictionary of measured output states and counts.
-
-        Returns:
-
-            float : The calculated expectation value.
-
-        """
-        expectation = 0
-        n_counts = 0
-        for state, counts in results.items():
-            n_counts += counts
-            # Adjust multiplier to account for variation in eigenvalues
-            multiplier = 1
-            for j, gate in enumerate(measurement.split(",")):
-                if gate == "I" or state[2 * j : 2 * j + 2] == State([1, 0]):
-                    multiplier *= 1
-                elif state[2 * j : 2 * j + 2] == State([0, 1]):
-                    multiplier *= -1
-                else:
-                    msg = (
-                        f"An invalid state {state[2 * j : 2 * j + 2]} was found"
-                        " in the results. This does not correspond to a valid "
-                        "value for dual-rail encoded qubits."
-                    )
-                    raise ValueError(msg)
-            expectation += multiplier * counts
-        return expectation / n_counts
